move game models into separate files
