+++ src/models/mod.rs
@@ -1,1 +1,4 @@
 pub mod game;
+pub mod leg;
+pub mod set;
+pub mod throw;
